pyproject.toml
[project]
name = "nita-bot"
version = "0.1.0"
description = "A discord bot for Mario Kart World"
authors = [
    {name = "Yumax-panda"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "dependency-injector (>=4.47.1,<5.0.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"

[tool.ruff.lint]
# ref: https://docs.astral.sh/ruff/linter/
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.lint.flake8-type-checking]
quote-annotations = true